Moved roots preprocessing algorithm to polyroots.py
