+++ sympy/polys/rootoftools.py
@@ -4,7 +4,7 @@ from sympy.core import (
     S, Basic, Expr, Integer, Real, I, Add, Lambda, symbols,
 )
 
-from sympy.polys.polytools import Poly, gcd_list, factor
+from sympy.polys.polytools import Poly, factor
 
 from sympy.polys.rootisolation import (
     dup_isolate_complex_roots_sqf,
@@ -12,7 +12,8 @@ from sympy.polys.rootisolation import (
 )
 
 from sympy.polys.polyroots import (
-    roots_linear, roots_quadratic, roots_binomial,
+    roots_linear, roots_quadratic,
+    roots_binomial, preprocess_roots,
 )
 
 from sympy.polys.rationaltools import (
@@ -27,12 +28,9 @@ from sympy.polys.polyerrors import (
     MultivariatePolynomialError,
     GeneratorsNeeded,
     PolynomialError,
-    DomainError,
 )
 
 from sympy.polys.domains import QQ
-
-from sympy.ntheory import divisors
 
 from sympy.mpmath import (
     mp, mpf, mpc, mpi, findroot,
@@ -277,107 +275,6 @@ def _rootof_data(poly, indices):
                 if index >= real_count:
                     yield _rootof_complexes_index(complexes, index-real_count)
 
-def integer_basis(poly):
-    monoms, coeffs = zip(*poly.terms())
-
-    monoms, = zip(*monoms)
-    coeffs = map(abs, coeffs)
-
-    if coeffs[0] < coeffs[-1]:
-        coeffs = list(reversed(coeffs))
-    else:
-        return None
-
-    monoms = monoms[:-1]
-    coeffs = coeffs[:-1]
-
-    divs = reversed(divisors(gcd_list(coeffs))[1:])
-
-    try:
-        div = divs.next()
-    except StopIteration:
-        return None
-
-    while True:
-        for monom, coeff in zip(monoms, coeffs):
-            if coeff % div**monom != 0:
-                try:
-                    div = divs.next()
-                except StopIteration:
-                    return None
-                else:
-                    break
-        else:
-            return div
-
-def _rootof_preprocess(poly):
-    """Try to get rid of symbolic coefficients from ``poly``. """
-    coeff = S.One
-
-    try:
-        _, poly = poly.clear_denoms(convert=True)
-    except DomainError:
-        return coeff, poly
-
-    poly = poly.primitive()[1]
-    poly = poly.retract()
-
-    if poly.get_domain().is_Poly:
-        poly = poly.inject()
-
-        strips = zip(*poly.monoms())
-        gens = list(poly.gens[1:])
-
-        base, strips = strips[0], strips[1:]
-
-        for gen, strip in zip(gens, strips):
-            reverse = False
-
-            if strip[0] < strip[-1]:
-                strip = reversed(strip)
-                reverse = True
-
-            ratio = None
-
-            for a, b in zip(base, strip):
-                if not a and not b:
-                    continue
-                elif not a or not b:
-                    break
-                elif b % a != 0:
-                    break
-                else:
-                    _ratio = b // a
-
-                    if ratio is None:
-                        ratio = _ratio
-                    elif ratio != _ratio:
-                        break
-            else:
-                if reverse:
-                    ratio = -ratio
-
-                poly = poly.eval(gen, 1)
-                coeff *= gen**(-ratio)
-                gens.remove(gen)
-
-        if gens:
-            poly = poly.eject(*gens)
-
-    if poly.is_univariate and poly.get_domain().is_ZZ:
-        basis = integer_basis(poly)
-
-        if basis is not None:
-            n = poly.degree()
-
-            def func((k,), coeff):
-                return coeff//basis**(n-k)
-
-            poly = poly.termwise(func)
-            coeff *= basis
-
-    return coeff, poly
-
 class RootOf(Expr):
     """Represents ``k``-th root of a univariate polynomial. """
 
@@ -430,7 +327,7 @@ class RootOf(Expr):
             else:
                 result = [ root for root in roots if root.is_real ]
         else:
-            coeff, poly = _rootof_preprocess(poly)
+            coeff, poly = preprocess_roots(poly)
             dom = poly.get_domain()
 
             if not dom.is_ZZ:
@@ -646,7 +543,7 @@ class RootSum(Expr):
     def _transform(cls, expr, x):
         """Transform an expression to a polynomial. """
         poly = Poly(expr, x, greedy=False)
-        return _rootof_preprocess(poly)
+        return preprocess_roots(poly)
 
     @classmethod
     def _is_func_rational(cls, poly, func):
